requirements.txt
flask==2.0.3
opencv-python==4.7.0
Pillow==9.2.0
numpy==1.23.1
firebase-admin==6.2.0
